requirements.txt
psycopg2==2.9.1
psycopg2-binary==2.9.9
Django==5.0.2
django-ninja==1.1.0
django-ninja-extra==0.20.7
django-ninja-jwt==5.3.1
ninja-schema==0.13.6
django-cors-headers==4.4.0
python-decouple==3.8